pyproject.toml
[project]
name = "pymcr-torch-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Brandon T. P. Saint-John", email = "saint-john@lbl.gov" }]
requires-python = ">=3.11"
dependencies = [
    "graphviz>=0.20.3",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "polars>=1.14.0",
    "pymcr>=0.5.1",
    "pyro-ppl>=1.9.1",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "torchinfo>=1.8.0",
    "torchtyping>=0.1.5",
]

[project.scripts]
pymcr-torch-demo = "pymcr_torch_demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
